feat: Add Sidebar component and navigation links
This commit adds a new component called Sidebar, which will be used for navigation in the application. It also includes an array of navigation links in the constants file.

The Sidebar component renders a set of icons based on the provided navigation links. Each icon can be clicked to navigate to the corresponding link. The active icon is highlighted with a different background color.

This change improves the overall user experience by providing a consistent and intuitive navigation system.