feat: Add createUser function to StateContextProvider
This commit adds the createUser function to the StateContextProvider component in the src/context/index.jsx file. The function is responsible for creating a new user in the database based on the provided userData. This change enhances the state management capabilities of the application by providing a method to add new users.