feat: Add edit mode and delete functionality to ColumnContainer component
This commit adds the edit mode and delete functionality to the ColumnContainer component. It introduces a state variable, editMode, to track whether the column is being edited. When the column title is clicked, the edit mode is enabled, allowing the user to update the column title. Additionally, a delete button is added to delete the column. These enhancements improve the usability and functionality of the kanban board.